+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 1.0.dev
 
 - Add .gitkeep to .gitignore.
+- Add scripts directory.
+- Change to `/vagrant` on login.
 
 ## Version 1.0